requirements.txt
alembic==1.7.4
anytree==2.8.0
apparmor==3.0.3
appdirs==1.4.4
argcomplete==1.12.1
Authlib==0.15.5
autopep8==1.6.0
Babel==2.9.1
bcrypt==3.2.0
Beaker==1.11.0
beautifulsoup4==4.10.0
bidict==0.21.4
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
btrfsutil==5.15.1
CacheControl==0.12.6
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==7.1.2
colorama==0.4.4
contextlib2==0.6.0.post1
cryptography==3.4.8
cupshelpers==1.0
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.3
distro==1.6.0
dnspython==1.16.0
docutils==0.17.1
email-validator==1.1.3
evdev==1.4.0
eventlet==0.31.0
extras==1.0.0
fixtures==3.0.0
Flask==1.1.2
Flask-Assets==2.0
Flask-BabelEx==0.9.4
Flask-Compress==1.8.0
Flask-Gravatar==0.5.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-MySQLdb==0.2.0
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security-Too==4.0.1
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
future==0.18.2
gevent==21.8.0
Glances==3.2.4
greenlet==1.1.2
gssapi==1.6.14
gunicorn==20.1.0
html5lib==1.1
httpagentparser==1.9.1
idna==2.10
importlib-metadata==4.8.1
isc==2.0
itsdangerous==1.1.0
Jinja2==2.11.3
ldap3==2.9
lensfun==0.3.95
LibAppArmor==3.0.3
libfdt==1.6.1
louis==3.19.0
lutris==0.5.9.1
lxml==4.6.4
Mako==1.1.6
mallard-ducktype==1.0.2
Markdown==3.3.6
MarkupSafe==2.0.1
meson==0.60.2
more-itertools==8.10.0
msgpack==1.0.2
mysqlclient==2.1.0
netsnmp-python==1.0a1
ordered-set==4.0.2
packaging==20.9
paramiko==2.7.2
passlib==1.7.4
pbr==5.7.0
pep517==0.12.0
pgadmin4==5.7
Pillow==8.4.0
ply==3.11
powerline-status==2.8.2
progress==1.6
protobuf==3.17.3
psutil==5.8.0
psycopg2==2.8.4
psycopg2-binary==2.8.6
pwquality==1.4.4
pyasn1==0.4.8
pycairo==1.20.1
pycodestyle==2.8.0
pycparser==2.21
pycups==2.0.1
pycurl==7.44.1
Pygments==2.10.0
PyGObject==3.42.0
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-editor==1.0.4
python-engineio==4.3.0
python-socketio==5.5.0
pytz==2021.3
pywal==3.3.1
pyxdg==0.27
PyYAML==5.4.1
ranger-fm==1.9.3
Reflector==2021.7.8.19.52.52
reportlab==3.6.2
requests==2.25.1
resolvelib==0.5.5
retrying==1.3.3
s-tui==1.1.3
simplejson==3.17.6
six==1.16.0
soupsieve==2.2.1
speaklater==1.3
speaklater3==1.4
spotipy==2.7.1
spotui==0.1.19
SQLAlchemy==1.3.24
sqlparse==0.4.2
sshtunnel==0.4.0
systemd-python==234
sysv-ipc==1.1.0
team==1.0
termcolor==1.1.0
testtools==2.5.0
tldr==2.0.0
toml==0.10.2
tomli==1.2.2
ua-parser==0.10.0
urllib3==1.26.7
urwid==2.1.2
user-agents==2.2.0
webassets==2.0
webencodings==0.5.1
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
